9. Adding CSV Writer for Pairing Results
Today, we're adding CSV writer to our file as a way to output a CSV of all the results we'll get from checkPair.
We'll also take a look at money-legos and what that does, it is an extremely useful library that is heavily used.

Lastly, we'll start looking at ethers instead of web3.js because in this project, we will be using ethers instead of web3.js

https://web3js.readthedocs.io/en/v1.2.11/web3-utils.html#fromwei
https://money-legos.studydefi.com/

HOMEWORK: Compare and contrast the difference between index.js and index2.js