12. Introduction to Aave Flashloans
Today, we're going to introduce the code for Aave Flash Loans, which is really one of the biggest (if not the biggest)
lending platforms in DeFi.

Homework: Copy and Paste what I ask you to into your own local project (i.e. a lot of new interfaces) and read over the Aave documentation.
References:
https://github.com/Uniswap/v2-core/tree/master/contracts/interfaces
https://github.com/Uniswap
https://github.com/Uniswap/v2-periphery/blob/master/contracts/libraries/UniswapV2Library.sol
https://docs.aave.com/developers/guides/flash-loans#example-code