13. Set up routers for swaps with Aave
Today, we're writing some skeleton code to make exchanges on AaveV2FlashLoan. What this skeleton code will do is run the swaps.
This is going to be completely trial and error as we think about the architecture whilst coding.

Homework: Copy and Paste the new interface files into your own local project.