16. Refactoring Code Based on PatrickAlphaC's Flashloan Repo
So, today is a refactoring day. When I was developing this project, I looked through many resources and Patrick's repo was one that I
found very helpful. The link is below if you want to use it as an example. But you will get by with just following the changes in this
video.

Reference:
https://github.com/PatrickAlphaC/aave-flashloan-mix