22. Getting Started with Ethernal – Local Etherscan
Today, we're going to do some more refactoring with AaveV2FlashLoan and you will be introduced to Ethernal, which is used as a local etherscan.
Etherscan is important to help us see the state of our transactions, but that can't be easily seen in local blockchains in our computer – so Ethernal helps with that.

Homework:
- In AaveV2FlashLoan.sol, I missed something that may cause the file to not compile. What is that something?

References:
NPM Package: https://www.npmjs.com/package/hardhat-ethernal
Docs: https://doc.tryethernal.com/