24. Adding Fees to the Frontend
We don't want to make calls to the contract if we don't have a profitable trade on our hands, so let's check on the frontend if we do – and only then, we call the smart contract.
Let's also talk about the use of decimals and how we can't use decimals in solidity and how we get around that.