27. Adding Tests to AaveV2FlashLoan
The funny thing about our DApp is that the best way to test that it works is through TESTS. Which is great because we create tests for code that we've written to ensure that both our code is valid and that what we want to do actually works.

Testing is a very important part of programming and we should always strive to write tests for everything we write.

References:

https://hardhat.org/hardhat-network/guides/mainnet-forking.html
https://hardhat.org/hardhat-network/reference/#hardhat-impersonateaccount
https://hardhat.org/tutorial/testing-contracts.html