28. Adding the Swap Functions
Today, we add Swap functions for Kyber, Uniswap, and Sushiswap that will also keep track of all the fees and whether the amount we receive back for asset 0 is actually more than the cost of the arbitrage to make sure we profit.

We'll add to the tests and talk about which parts of the app you have to customize for yourself at the end of this series because this video series gives you more of a starting point for flash loan arbitrage rather than an end point.

Best of luck!

References:
https://docs.uniswap.org/protocol/V2/reference/smart-contracts/router-02#swapexacttokensfortokens
https://docs.kyberswap.com/Legacy/api-abi/core-smart-contracts/api_abi-isimplekyberproxy#swaptokentotoken