30. Finishing Touches. The Finale.
That's a wrap people. We're going to put on some finishing touches and some quick refactoring and see how our price bot works.
Like I said before, this is not the end point, there's still a lot of optimizations for this price bot to work properly and to
make money for you in the field. This whole entire project was meant to give you a "STARTING POINT" for you to build your own
custom gunda- price bot that you can be proud of and that will make you millions!

Until next time!